Move websockets tests to separate file.
