+++ Raven.Tests.Core/ChangesApi/Subscribing.cs
@@ -1,13 +1,10 @@
-﻿using System.Net.WebSockets;
-using System.Text;
-using System.Threading;
+﻿using System.Threading;
 using Raven.Abstractions.Data;
 using Raven.Json.Linq;
 using Raven.Tests.Core.Replication;
 using Raven.Tests.Core.Utils.Entities;
 using Raven.Tests.Core.Utils.Indexes;
 using System;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace Raven.Tests.Core.ChangesApi
@@ -184,27 +181,6 @@ namespace Raven.Tests.Core.ChangesApi
                 source.Replication.WaitAsync(eTag, replicas: 1).Wait();
 
                 WaitUntilOutput("conflict");
-            }
-        }
-
-        [Fact]
-        public async Task Can_connect_via_websockets_and_receive_heartbeat()
-        {
-            using (var store = GetDocumentStore())
-            {
-                using (var clientWebSocket = new ClientWebSocket())
-                {
-                    string url = store.Url.Replace("http:", "ws:");
-                    url = url + "/changes/websocket?id=" + Guid.NewGuid();
-                    await clientWebSocket.ConnectAsync(new Uri(url), CancellationToken.None);
-
-                    var buffer = new byte[1024];
-                    WebSocketReceiveResult result =
-                        await clientWebSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-                    var message = Encoding.UTF8.GetString(buffer, 0, result.Count);
-
-                    Assert.Contains("Heartbeat", message);
-                }
             }
         }
     }
