+++ Raven.Tests.Core/Raven.Tests.Core.csproj
@@ -102,6 +102,7 @@
   <ItemGroup>
     <Compile Include="ChangesApi\ImplementingChangesClient.cs" />
     <Compile Include="ChangesApi\Subscribing.cs" />
+    <Compile Include="ChangesApi\WebsocketsTests.cs" />
     <Compile Include="Commands\Batches.cs" />
     <Compile Include="Commands\Querying.cs" />
     <Compile Include="Commands\Transformers.cs" />
